+++ src/test/java/com/neoproduction/factorialguesser/FactorialGuesserBruteForceTest.java
@@ -2,6 +2,8 @@ package com.neoproduction.factorialguesser;
 
 import org.junit.Test;
 
+import java.math.BigInteger;
+
 import static com.neoproduction.factorialguesser.FactorialGuesser.BAD_FACTORIAL;
 import static org.junit.Assert.*;
 
@@ -10,37 +12,37 @@ public class FactorialGuesserBruteForceTest {
     @Test
     public void getNFromFactorial_Negative() {
         FactorialGuesser factorialGuesser = new FactorialGuesserBruteForce();
-        assertEquals(BAD_FACTORIAL, factorialGuesser.getNFromFactorial(-1));
+        assertEquals(BAD_FACTORIAL, factorialGuesser.getNFromFactorial(BigInteger.valueOf(-1)));
     }
 
     @Test
     public void getNFromFactorial_NoSuchFactorial() {
         FactorialGuesser factorialGuesser = new FactorialGuesserBruteForce();
-        assertEquals(BAD_FACTORIAL, factorialGuesser.getNFromFactorial(7));
+        assertEquals(BAD_FACTORIAL, factorialGuesser.getNFromFactorial(BigInteger.valueOf(7)));
     }
 
     @Test
     public void getNFromFactorial_1() {
         FactorialGuesser factorialGuesser = new FactorialGuesserBruteForce();
-        int result = factorialGuesser.getNFromFactorial(1);
+        int result = factorialGuesser.getNFromFactorial(BigInteger.valueOf(1));
         assertTrue(result == 1 || result == 0);
     }
 
     @Test
     public void getNFromFactorial_Small() {
         FactorialGuesser factorialGuesser = new FactorialGuesserBruteForce();
-        assertEquals(2, factorialGuesser.getNFromFactorial(2));
+        assertEquals(2, factorialGuesser.getNFromFactorial(BigInteger.valueOf(2)));
     }
 
     @Test
     public void getNFromFactorial_Medium() {
         FactorialGuesser factorialGuesser = new FactorialGuesserBruteForce();
-        assertEquals(7, factorialGuesser.getNFromFactorial(5040));
+        assertEquals(7, factorialGuesser.getNFromFactorial(BigInteger.valueOf(5040)));
     }
 
     @Test
     public void getNFromFactorial_Big() {
         FactorialGuesser factorialGuesser = new FactorialGuesserBruteForce();
-        assertEquals(12, factorialGuesser.getNFromFactorial(479001600L));
+        assertEquals(12, factorialGuesser.getNFromFactorial(BigInteger.valueOf(479001600L)));
     }
 }
